requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.4
django-ckeditor==6.7.2
django-debug-toolbar==5.2.0
django-js-asset==3.1.2
django-registration==5.2.1
django-widget-tweaks==1.5.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
